move ServiceHost to mythfrontend
