+++ mythtv/libs/libmythupnp/CMakeLists.txt
@@ -17,7 +17,6 @@ set(HEADERS_TO_INSTALL1
     mmembuf.h
     mythxmlclient.h
     serverSideScripting.h
-    servicehost.h
     soapclient.h
     ssdp.h
     ssdpcache.h
@@ -33,8 +32,7 @@ set(HEADERS_TO_INSTALL1
     upnptasknotify.h
     upnptasksearch.h
     upnputil.h
-    wsdl.h
-    xsd.h)
+    )
 
 set(HEADERS_TO_INSTALL2
     serializers/jsonSerializer.h serializers/serializer.h
@@ -72,8 +70,6 @@ add_library(
   serializers/xmlplistSerializer.h
   serializers/xmlSerializer.cpp
   serializers/xmlSerializer.h
-  servicehost.cpp
-  servicehost.h
   soapclient.cpp
   soapclient.h
   ssdp.cpp
@@ -113,10 +109,7 @@ add_library(
   websocket.h
   websocket_extensions/websocket_mythevent.cpp
   websocket_extensions/websocket_mythevent.h
-  wsdl.cpp
-  wsdl.h
-  xsd.cpp
-  xsd.h)
+  )
 
 #
 # All remaining target information
