+++ mythtv/programs/mythfrontend/serviceHosts/servicehost.h
@@ -23,7 +23,6 @@
 
 #include "libmythupnp/httprequest.h"
 #include "libmythupnp/httpserver.h"
-#include "libmythupnp/upnpexp.h"
 
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
@@ -33,7 +32,7 @@
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
 
-class UPNP_PUBLIC MethodInfo
+class MethodInfo
 {
     public:
 
@@ -62,7 +61,7 @@ using MetaInfoMap = QMap< QString, MethodInfo >;
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
 
-class UPNP_PUBLIC ServiceHost : public HttpServerExtension
+class ServiceHost : public HttpServerExtension
 {
     protected:
