+++ mythtv/programs/mythfrontend/serviceHosts/wsdl.h
@@ -19,9 +19,8 @@
 #include <QMetaObject>
 #include <QString>
 
-#include "libmythupnp/upnpexp.h"
-#include "libmythupnp/servicehost.h"
-#include "libmythupnp/xsd.h"
+#include "servicehost.h"
+#include "xsd.h"
 
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
@@ -31,7 +30,7 @@
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
 
-class UPNP_PUBLIC Wsdl : public QDomDocument
+class Wsdl : public QDomDocument
 {
     protected:
         ServiceHost             *m_pServiceHost {nullptr};
