+++ mythtv/programs/mythfrontend/serviceHosts/xsd.h
@@ -19,7 +19,6 @@
 #include <QDomDocument>
 
 #include "libmythupnp/httprequest.h"
-#include "libmythupnp/upnpexp.h"
 
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
@@ -29,7 +28,7 @@
 //////////////////////////////////////////////////////////////////////////////
 //////////////////////////////////////////////////////////////////////////////
 
-class UPNP_PUBLIC Xsd : public QDomDocument
+class Xsd : public QDomDocument
 {
     protected:
